débuts fonction generate_sudoku
